Move braille offset converter helper out of braille module
